requirements.txt
Flask==2.1.2
python-chess
gunicorn==20.1.0
Werkzeug==2.1.2
